bml: swap preset loading order
We first load user presets and then sysem presets. This is in sync with the
other preset behaviour. If system wide version get updated we don't want to
shadow it. One remaining problem is that if the user overwrites a preset
entry this would only work if it gets a new name.